implemented updated delete and put method for transaction

+++ YourCard/src/main/java/com/ness/YourCard/controller/TransactionController.java
@@ -38,16 +38,28 @@ public UserService userService;
     @Transactional
     @PutMapping(value= "/UpdateTransaction/{TransactionId}")
     public String updateCustomer(@PathVariable Integer TransactionId, @RequestBody TransactionDTO transactionDTO) {
-        transactionService.updateTransactionDetails(TransactionId,transactionDTO);
-        return "transaction with id: " + TransactionId + " updated succesfully";
+       if(transactionService.checkIfTransactionPresent(TransactionId)) {
+           transactionService.updateTransactionDetails(TransactionId, transactionDTO);
+           return "transaction with id: " + TransactionId + " updated succesfully";
+       }
+       else
+           return "no such transaction with id: "+TransactionId+" is present";
     }
 
 
     @Transactional
     @DeleteMapping(value = "/DeleteTransaction/{TransactionId}")
      public String delTransaction(@PathVariable Integer TransactionId){
+        if(transactionService.checkIfTransactionPresent(TransactionId)){
             transactionService.deleteTransaction(TransactionId);
             return " Transaction with id" + TransactionId + " deleted successfully";
+        }
+        else {
+            return "No such transaction present with id: "+ TransactionId;
+
+        }
+
+
         }
